Centralized carveout location to amp-ad.js.
