+++ examples/sticky.ads.amp.html
@@ -161,6 +161,7 @@
   <amp-sticky-ad layout="nodisplay">
     <amp-ad width=320 height=50
       type="doubleclick"
+      data-multi-size="300x50"
       data-slot="/4119129/mobile_ad_banner">
   </amp-ad>
